# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: University Management System (UMS) API
  version: 1.0.0
  description: |
    REST API for the University Management System built with Laravel 11 (Modular Monolith).

    ## Authentication
    All endpoints (except Login and Refresh) require a JWT Bearer token in the `Authorization` header.

    ## Roles & Permissions
    - **Admin** — Full access (all 32 permissions)
    - **Registration Staff** — Courses, sections, classrooms, student read, enrollment read
    - **Student** — Enroll/drop, view courses/sections (filtered by department), timetable, academic read

    ## Modules
    - **Identity** — Authentication, users, roles, permissions
    - **Academic** — Universities, colleges, departments, majors
    - **Registration** — Courses, sections, classrooms, schedules
    - **Student** — Students, enrollments, timetable
  contact:
    name: UMS Development Team
servers:
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/majors:
    get:
      tags: [Majors]
      summary: List majors
      description: "Permission required: `academic.read`"
      operationId: listMajors
      parameters:
        - name: department_id
          in: query
          schema:
            type: integer
        - $ref: '#/components/parameters/search'
        - $ref: '#/components/parameters/per_page'
      responses:
        '200':
          description: Paginated list of majors
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Major'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
    post:
      tags: [Majors]
      summary: Create major
      description: "Permission required: `academic.create`"
      operationId: createMajor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [department_id, name, code, total_credits]
              properties:
                department_id:
                  type: integer
                name:
                  type: string
                  maxLength: 200
                code:
                  type: string
                  maxLength: 20
                total_credits:
                  type: integer
                  minimum: 0
      responses:
        '201':
          description: Major created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Major'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/majors/{majorId}:
    get:
      tags: [Majors]
      summary: Get major by ID
      description: "Permission required: `academic.read`"
      operationId: getMajor
      parameters:
        - name: majorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Major details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Major'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Majors]
      summary: Update major
      description: "Permission required: `academic.update`"
      operationId: updateMajor
      parameters:
        - name: majorId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                department_id:
                  type: integer
                name:
                  type: string
                  maxLength: 200
                code:
                  type: string
                  maxLength: 20
                total_credits:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: Major updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Major'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Majors]
      summary: Delete major
      description: "Permission required: `academic.delete`"
      operationId: deleteMajor
      parameters:
        - name: majorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Major deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # COURSES
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from POST /api/auth/login
  parameters:
    search:
      name: search
      in: query
      description: Search keyword
      schema:
        type: string
    per_page:
      name: per_page
      in: query
      description: Items per page (default 15)
      schema:
        type: integer
        default: 15
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: The given data was invalid.
              errors:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Resource not found
    Message:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully
    PaginationMeta:
      type: object
      properties:
        current_page:
          type: integer
        last_page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        from:
          type: integer
          nullable: true
        to:
          type: integer
          nullable: true
...
    Major:
      type: object
      properties:
        major_id:
          type: integer
        department_id:
          type: integer
        name:
          type: string
        code:
          type: string
        total_credits:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...